update regex docs with more info about source gen
