+++ .openpublishing.redirection.standard.json
@@ -213,7 +213,11 @@
         },
         {
             "source_path_from_root": "/docs/standard/base-types/thread-safety.md",
-            "redirect_url": "/dotnet/standard/base-types/thread-safety-in-regular-expressions"
+            "redirect_url": "/dotnet/standard/base-types/best-practices-regex"
+        },
+        {
+            "source_path_from_root": "/docs/standard/base-types/thread-safety-in-regular-expressions.md",
+            "redirect_url": "/dotnet/standard/base-types/best-practices-regex"
         },
         {
             "source_path_from_root": "/docs/standard/base-types/verify-format.md",
