+++ docs/standard/base-types/snippets/regex/csharp/Program.cs
@@ -1,0 +1,6 @@
+﻿namespace Project;
+internal class Program
+{
+    public static void Main()
+        => CompareExample.RunIt();
+}
