+++ docs/standard/base-types/snippets/regex/csharp/Project.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>disable</Nullable>
+    <StartupObject>Project.Program</StartupObject>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="Dreiser_TheFinancier.txt">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+</Project>
